Use Blueprints to refactor the file structure
