+++ app/app.py
@@ -12,6 +12,8 @@ from werkzeug.security import generate_password_hash, check_password_hash
 from app.insert_sample_data import insert_sample_data
 from app.routes.unit import unit_bp
 from app.routes.myschedule import myschedule_bp
+from app.routes.auth import auth_bp
+
 from app.models import db
 from flask_migrate import Migrate
 from datetime import datetime
@@ -23,7 +25,7 @@ from flask_wtf.csrf import CSRFProtect
 
 app = Flask(__name__)
 app.config.from_object(Config)
-
+app.register_blueprint(auth_bp)
 # Initialize db and migrate
 db.init_app(app)
 migrate = Migrate(app, db)
@@ -51,82 +53,6 @@ def home():
         return render_template("home.html", name=current_user.username)
     else:
         return render_template("home.html", name="Guest")
-
-@app.route("/login", methods=["GET", "POST"])
-def login():
-    form = LoginForm()
-    if form.validate_on_submit():
-        user = User.query.filter_by(username=form.username.data).first()
-        if user and check_password_hash(user.password_hash, form.password.data):
-            login_user(user)
-            return redirect(url_for("home"))
-        flash("Invalid credentials", "danger")
-
-    return render_template("login.html", form=form)
-
-
-@app.route("/register", methods=["GET", "POST"])
-def register():
-    form = RegisterForm()
-    if form.validate_on_submit():
-        existing_user = User.query.filter_by(username=form.username.data).first()
-        if existing_user:
-            flash("Username already taken", "danger")
-            return render_template("register.html", form=form)
-
-        hashed_pw = generate_password_hash(form.password.data)
-        new_user = User(username=form.username.data, password_hash=hashed_pw)
-        db.session.add(new_user)
-        db.session.commit()
-        flash("Registration successful. You can now log in.", "success")
-        return redirect(url_for("login"))
-
-    return render_template("register.html", form=form)
-
-@app.route("/resetpw", methods=["GET", "POST"])
-def resetpw():
-    if request.method == "POST":
-        username = request.form.get("username")
-        user = User.query.filter_by(username=username).first()
-        if user:
-            return redirect(url_for("resetpw_username", username=username))
-        else:
-            flash("Username not found.", "danger")
-            return render_template("resetpw.html")
-    return render_template("resetpw.html")
-
-@app.route("/resetpw/<username>", methods=["GET", "POST"])
-def resetpw_username(username):
-    user = User.query.filter_by(username=username).first()
-    if not user:
-        flash("Invalid username.", "danger")
-        return redirect(url_for("resetpw"))
-
-    if request.method == "POST":
-        new_password = request.form.get("new_password")
-        confirm_password = request.form.get("confirm_password")
-
-        if new_password != confirm_password:
-            flash("Passwords do not match.", "danger")
-            return render_template("resetpw_form.html", username=username)
-
-        user.password_hash = generate_password_hash(new_password)
-        db.session.commit()
-        flash("Password reset successful. You can now log in.", "success")
-        return redirect(url_for("login"))
-
-    return render_template("resetpw_form.html", username=username)
-
-@app.route("/dashboard")
-@login_required
-def dashboard():
-    return render_template("dashboard.html", name=current_user.username)
-
-@app.route("/logout")
-@login_required
-def logout():
-    logout_user()
-    return redirect(url_for("home"))
 
 @app.route('/My_Schedule')
 @login_required
