+++ app/routes/auth.py
@@ -1,0 +1,72 @@
+# app/routes/auth.py
+
+from flask import Blueprint, render_template, request, redirect, url_for, flash
+from flask_login import login_user, logout_user, current_user
+from werkzeug.security import generate_password_hash, check_password_hash
+from app.models import User, db
+from app.forms import LoginForm, RegisterForm
+from app.forms import ResetPasswordForm
+
+auth_bp = Blueprint('auth', __name__)
+
+@auth_bp.route("/login", methods=["GET", "POST"])
+def login():
+    form = LoginForm()
+    if form.validate_on_submit():
+        user = User.query.filter_by(username=form.username.data).first()
+        if user and check_password_hash(user.password_hash, form.password.data):
+            login_user(user)
+            return redirect(url_for("home"))
+        flash("Invalid credentials", "danger")
+    return render_template("login.html", form=form)
+
+@auth_bp.route("/register", methods=["GET", "POST"])
+def register():
+    form = RegisterForm()
+    if form.validate_on_submit():
+        existing_user = User.query.filter_by(username=form.username.data).first()
+        if existing_user:
+            flash("Username already taken", "danger")
+            return render_template("register.html", form=form)
+
+        hashed_pw = generate_password_hash(form.password.data)
+        new_user = User(username=form.username.data, password_hash=hashed_pw)
+        db.session.add(new_user)
+        db.session.commit()
+        flash("Registration successful. You can now log in.", "success")
+        return redirect(url_for("auth.login"))
+    return render_template("register.html", form=form)
+
+@auth_bp.route("/logout")
+def logout():
+    logout_user()
+    return redirect(url_for("home"))
+
+@auth_bp.route("/resetpw", methods=["GET", "POST"])
+def resetpw():
+    if request.method == "POST":
+        username = request.form.get("username")
+        user = User.query.filter_by(username=username).first()
+        if user:
+            return redirect(url_for("auth.resetpw_username", username=username))
+        else:
+            flash("Username not found.", "danger")
+            return render_template("resetpw.html")
+    return render_template("resetpw.html")
+
+
+@auth_bp.route("/resetpw/<username>", methods=["GET", "POST"])
+def resetpw_username(username):
+    user = User.query.filter_by(username=username).first()
+    if not user:
+        flash("Invalid username.", "danger")
+        return redirect(url_for("auth.resetpw"))
+
+    form = ResetPasswordForm()
+    if form.validate_on_submit():
+        user.password_hash = generate_password_hash(form.new_password.data)
+        db.session.commit()
+        flash("Password reset successful. You can now log in.", "success")
+        return redirect(url_for("auth.login"))
+
+    return render_template("resetpw_form.html", username=username, form=form)
