+++ app/templates/base.html
@@ -95,17 +95,17 @@
               </li>
               <li><hr class="dropdown-divider" /></li>
               <li>
-                <a class="dropdown-item" href="{{ url_for('logout') }}"
+                <a class="dropdown-item" href="{{ url_for('auth.logout') }}"
                   ><i class="bi bi-box-arrow-right me-2"></i>Logout</a
                 >
               </li>
             </ul>
           </div>
           {% else %}
-          <a href="{{ url_for('login') }}" class="btn btn-outline-primary me-2"
+          <a href="{{ url_for('auth.login') }}" class="btn btn-outline-primary me-2"
             >Login</a
           >
-          <a href="{{ url_for('register') }}" class="btn btn-primary"
+          <a href="{{ url_for('auth.register') }}" class="btn btn-primary"
             >Sign Up</a
           >
           {% endif %}
