--- app/templates/dashboard.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>User Dashboard</title>
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.5/dist/css/bootstrap.min.css" rel="stylesheet">
-</head>
-<body>
-  <div class="container mt-5">
-    <h1 class="mb-4">Welcome, {{ name }}!</h1>
-    <p>This is your personal dashboard. Manage your courses and schedule here.</p>
-    <a href="{{ url_for('logout') }}" class="btn btn-danger">Logout</a>
-  </div>
-</body>
-</html>
