+++ app/templates/home.html
@@ -34,10 +34,10 @@
             <span class="me-2">{{ current_user.username }}</span>
             <img src="{{ url_for('static', filename='image/avatar.png') }}" alt="Avatar" class="rounded-circle" width="32" height="32">
           </a>
-          <a href="{{ url_for('logout') }}" class="btn btn-outline-primary ms-3">Logout</a>
+          <a href="{{ url_for('auth.logout') }}" class="btn btn-outline-primary ms-3">Logout</a>
         {% else %}
-          <a href="{{ url_for('login') }}" class="btn btn-outline-primary me-2">Login</a>
-          <a href="{{ url_for('register') }}" class="btn btn-primary">Sign Up</a>
+          <a href="{{ url_for('auth.login') }}" class="btn btn-outline-primary me-2">Login</a>
+          <a href="{{ url_for('auth.register') }}" class="btn btn-primary">Sign Up</a>
         {% endif %}
       </div>
