+++ app/templates/register.html
@@ -44,26 +44,6 @@
             {% endif %}
             {% endwith %}
 
-            <!-- <form method="POST" action="{{ url_for('register') }}">
-              <div class="mb-3">
-                <input type="text" name="username" class="form-control" placeholder="Username" required />
-              </div>
-              <div class="mb-3">
-                <input type="password" name="password" class="form-control" placeholder="Password" required />
-              </div>
-              <div class="mb-3">
-                <input type="password" name="confirm_password" class="form-control" placeholder="Confirm Password"
-                  required />
-              </div>
-                            <div class="d-grid">
-                <button type="submit" class="btn btn-primary-custom w-100">Sign Up</button>
-              </div>
-            </form> -->
-
-            <!-- 
-            <div class="text-center mt-3">
-              <a href="{{ url_for('login') }}" class="text-primary">Go back to the login page</a> -->
-
             <form method="POST">
               {{ form.hidden_tag() }}
 
@@ -96,7 +76,7 @@
               </div>
             </form>
 
-            <p class="mt-3">Already have an account? <a href="{{ url_for('login') }}">Login here</a>.</p>
+            <p class="mt-3">Already have an account? <a href="{{ url_for('auth.login') }}">Login here</a>.</p>
           </div>
         </div>
       </div>
