+++ app/templates/resetpw.html
@@ -37,7 +37,7 @@
             <!-- <div class="alert alert-info" style="display: none;"></div> -->
             <div class="alert alert-info d-none" role="alert" id="statusMessage"></div>
 
-            <form method="POST" action="{{ url_for('resetpw') }}" id="resetPasswordForm">
+            <form method="POST" action="{{ url_for('auth.resetpw') }}" id="resetPasswordForm">
               <!-- <div class="mb-3">
                 <input type="email" name="email" class="form-control" placeholder="Enter your registered email" required id="emailInput">
               </div> -->
@@ -50,7 +50,7 @@
             </form>
 
             <div class="text-center mt-3">
-              <a href="{{ url_for('login') }}" class="text-primary">Back to Login</a>
+              <a href="{{ url_for('auth.login') }}" class="text-primary">Back to Login</a>
             </div>
           </div>
         </div>
