+++ app/templates/resetpw_form.html
@@ -39,20 +39,6 @@
             <!-- <div class="alert alert-info" style="display: none;"></div> -->
             <div class="alert alert-info d-none" role="alert" id="statusMessage"></div>
 
-
-            <!-- <form method="POST" action="{{ url_for('resetpw_username', username=username) }}">
-                <div class="mb-3">
-                  <input type="password" name="new_password" class="form-control" placeholder="Password" required />
-                </div>
-                <div class="mb-3">
-                  <input type="password" name="confirm_password" class="form-control" placeholder="Confirm Password" required />
-                </div>
-                <div class="d-grid">
-                  <button type="submit" class="btn btn-primary-custom w-100">Submit</button>
-                </div>
-              </form> -->
-
-
             <form method="POST">
               {{ form.hidden_tag() }}
 
@@ -78,7 +64,7 @@
             </form>
 
             <div class="text-center mt-3">
-              <a href="{{ url_for('login') }}" class="text-primary">Back to Login</a>
+              <a href="{{ url_for('auth.login') }}" class="text-primary">Back to Login</a>
             </div>
           </div>
         </div>
